Update prose tests.
